requirements.txt
elevenlabs
langgraph
langchain
python-dotenv
langchain-groq
pydantic
fastapi
fastapi[all]
uvicorn[standard]